resolve clang-tidy errors

+++ src/pcms/transfer/load_vector_integrator.cpp
@@ -314,7 +314,7 @@ Kokkos::View<MeshField::Real*> buildLoadVector(
 
   // TODO: Make it generalised; hardcoded for liner 2D
   Kokkos::View<MeshField::Real*> elmLoadVector("elmLoadVector",
-                                               target_mesh.nelems() * 3);
+                                               static_cast<long>(target_mesh.nelems()) * 3);
   Kokkos::parallel_for(
     "calculate load vector", target_mesh.nelems(),
     KOKKOS_LAMBDA(const int& elm) {
